+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/dto/TagDto.java
@@ -23,4 +23,8 @@ public class TagDto {
 
     @JsonProperty(value = "tag_description", index = 2)
     private String tagDescription;
+
+    @JsonProperty("language_id")
+    private UUID languageId;
+
 }